# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gupta Enterprises E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: inStock
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{productId}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product details
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (Admin)
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (Admin)
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # CATEGORIES
...
  /reviews/product/{productId}:
    get:
      operationId: getProductReviews
      tags: [reviews]
      summary: Get reviews for a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Add a review (requires delivered order)
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{reviewId}:
    put:
      operationId: updateReview
      tags: [reviews]
      summary: Update a review
      parameters:
        - name: reviewId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReviewBody"
      responses:
        "200":
          description: Updated review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # USERS (Admin)
...
  /tracking/{orderId}:
    get:
      operationId: getOrderTracking
      tags: [tracking]
      summary: Get live tracking for an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tracking info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackingInfo"
  /tracking/{orderId}/location:
    put:
      operationId: updateDeliveryLocation
      tags: [tracking]
      summary: Update delivery agent location (Delivery Agent)
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLocationBody"
      responses:
        "200":
          description: Location updated

  # NOTIFICATIONS
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        firebaseUid:
          type: string
        email:
          type: string
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [USER, ADMIN, DELIVERY_AGENT]
        photoUrl:
          type: string
        addresses:
          type: array
          items:
            $ref: "#/components/schemas/Address"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - role
        - createdAt
    Address:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        street:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        lat:
          type: number
        lng:
          type: number
        isDefault:
          type: boolean
      required:
        - id
        - street
        - city
        - state
        - pincode
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        categoryId:
          type: string
        category:
          $ref: "#/components/schemas/Category"
        actualPrice:
          type: number
        sellingPrice:
          type: number
        discount:
          type: number
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - actualPrice
        - sellingPrice
        - stock
        - images
        - isActive
        - createdAt
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - products
        - total
        - page
        - limit
        - totalPages
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
        productCount:
          type: integer
      required:
        - id
        - name
        - slug
...
    Review:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        userId:
          type: string
        user:
          $ref: "#/components/schemas/User"
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: string
        body:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - productId
        - userId
        - rating
        - createdAt
    TrackingInfo:
      type: object
      properties:
        orderId:
          type: string
        status:
          type: string
        agentName:
          type: string
        agentPhone:
          type: string
        currentLat:
          type: number
        currentLng:
          type: number
        destinationLat:
          type: number
        destinationLng:
          type: number
        estimatedDelivery:
          type: string
          format: date-time
        history:
          type: array
          items:
            $ref: "#/components/schemas/TrackingEvent"
      required:
        - orderId
        - status
    TrackingEvent:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        lat:
          type: number
        lng:
          type: number
      required:
        - status
        - message
        - timestamp
...
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        categoryId:
          type: string
        actualPrice:
          type: number
        sellingPrice:
          type: number
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
      required:
        - name
        - description
        - actualPrice
        - sellingPrice
        - stock
        - images
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        categoryId:
          type: string
        actualPrice:
          type: number
        sellingPrice:
          type: number
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
        isActive:
          type: boolean
...
    CreateReviewBody:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: string
        body:
          type: string
        orderId:
          type: string
      required:
        - rating
        - orderId
    UpdateReviewBody:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: string
        body:
          type: string
...
    UpdateLocationBody:
      type: object
      properties:
        lat:
          type: number
        lng:
          type: number
      required:
        - lat
        - lng